+++ lru_example_test.go
@@ -299,7 +299,7 @@ func ExamplePoolLRU_Upsert() {
 		Name:  "justpranavrs",
 		Email: "iliketlru@gmail.com",
 	})
-	if state == core.AddNoEvict {
+	if state == core.UpsertAddNoEviction {
 		fmt.Println("[UPSERT] : Add without Eviction")
 	}
 
@@ -313,7 +313,7 @@ func ExamplePoolLRU_Upsert() {
 		Name:  "welcometotlru",
 		Email: "welcometotlru@gmail.com",
 	})
-	if state == core.AddNoEvict {
+	if state == core.UpsertAddNoEviction {
 		fmt.Println("[UPSERT] : Add without Eviction")
 	}
 
@@ -324,7 +324,7 @@ func ExamplePoolLRU_Upsert() {
 		Name:  "justpranavrs",
 		Email: "tlruiscool@gmail.com",
 	})
-	if state == core.AddOnEvict {
+	if state == core.UpsertAddWithEviction {
 		fmt.Println("[UPSERT] : Add on Eviction")
 		fmt.Printf("[UPSERT] Name : %v | Email : %v\n", val.Name, val.Email)
 	}
@@ -339,7 +339,7 @@ func ExamplePoolLRU_Upsert() {
 		Name:  "justpranavrs",
 		Email: "jprs-tlru@gmail.com",
 	})
-	if state == core.Replace {
+	if state == core.UpsertReplace {
 		fmt.Println("[UPSERT] : Value Replaced")
 		fmt.Printf("[UPSERT] Name : %v | Email : %v\n", val.Name, val.Email)
 	}