+++ tpool.go
@@ -55,14 +55,27 @@ type TPool[K comparable, V any] interface {
 
 // tPool is the internal sharding container used by [PoolTLRU] and [PoolTSLRU].
 // It handles routing via [mux.Mux] and maintains the shards.
+//
+// tPool does not implement pool because GoDoc doesn't recognize multiple
+// levels of struct embedding even though compiler can.
 type tPool[K comparable, V any, C core.TTLShard[K, V]] struct {
-	pool[K, V, C]
-	clock *clock.Clock
+	// capacity represents the maximum allocated space for the LRU cache.
+	capacity int
+
+	// shards is an array of the multiple instances of [core.Shard].
+	shards []C
+
+	// mux is the router for the shards in the shards array.
+	// It takes in a key K and outputs a hash [uint32]
+	mux mux.Mux[K]
+
+	// ttl clock and sliding ttl
+	clock   *clock.Clock
 	sliding bool
 }
 
 // assembleWithTTL creates a [tPool] instance with the provided arguments. It creates shards based on the createShard function.
-func assembleWithTTL[K comparable, V any, C core.TTLShard[K, V]](capacity int, nShards int, 
+func assembleWithTTL[K comparable, V any, C core.TTLShard[K, V]](capacity int, nShards int,
 	hash mux.Mux[K], clock *clock.Clock, sliding bool, createShard func(cap int) (C, error),
 ) (tPool[K, V, C], error) {
 	var zero tPool[K, V, C]
@@ -87,13 +100,11 @@ func assembleWithTTL[K comparable, V any, C core.TTLShard[K, V]](capacity int, n
 		}
 	}
 	return tPool[K, V, C]{
-		pool: pool[K, V, C]{
-			capacity: capacity,
-			shards: shards,
-			mux: hash,
-		},
-		clock: clock,
-		sliding: sliding,
+		capacity: capacity,
+		shards:   shards,
+		mux:      hash,
+		clock:    clock,
+		sliding:  sliding,
 	}, nil
 }
 
@@ -108,14 +119,14 @@ func (l *tPool[K, V, C]) Close() {
 // It also returns the remaining TTL in the key if it was found in the cache.
 func (l *tPool[K, V, C]) GetWithTTL(key K) (V, time.Duration, bool) {
 	shard := l.mux(key)
-	return l.pool.shards[shard].GetWithTTL(key)
+	return l.shards[shard].GetWithTTL(key)
 }
 
 // Check [PoolTLRU.Peek] on how Peek works.
 // It also returns the remaining TTL in the key if it was found in the cache.
 func (l *tPool[K, V, C]) PeekWithTTL(key K) (V, time.Duration, bool) {
 	shard := l.mux(key)
-	return l.pool.shards[shard].PeekWithTTL(key)
+	return l.shards[shard].PeekWithTTL(key)
 }
 
 // PutWithTTL adds a new value to the cache with the given key and the provided ttl value.
@@ -125,14 +136,14 @@ func (l *tPool[K, V, C]) PeekWithTTL(key K) (V, time.Duration, bool) {
 // Check [PoolTLRU.Put] on how Put works.
 func (l *tPool[K, V, C]) PutWithTTL(key K, value V, ttl time.Duration) {
 	shard := l.mux(key)
-	l.pool.shards[shard].PutWithTTL(key, value, ttl)
+	l.shards[shard].PutWithTTL(key, value, ttl)
 }
 
 // Refresh resets the TTL of an existing key using the default ttl.
 // It returns false if the key could not be found.
 func (l *tPool[K, V, C]) Refresh(key K) bool {
 	shard := l.mux(key)
-	return l.pool.shards[shard].Refresh(key)
+	return l.shards[shard].Refresh(key)
 }
 
 // SetTTL resets the TTL of an existing key using the provided ttl value.
@@ -141,13 +152,13 @@ func (l *tPool[K, V, C]) Refresh(key K) bool {
 // The ttl value is rounded off in terms of its internal clock ticks.
 func (l *tPool[K, V, C]) SetTTL(key K, ttl time.Duration) bool {
 	shard := l.mux(key)
-	return l.pool.shards[shard].SetTTL(key, ttl)
+	return l.shards[shard].SetTTL(key, ttl)
 }
 
 // TTL returns the remaining TTL for the key.
 func (l *tPool[K, V, C]) TTL(key K) (time.Duration, bool) {
 	shard := l.mux(key)
-	return l.pool.shards[shard].TTL(key)
+	return l.shards[shard].TTL(key)
 }
 
 // UpsertWithTTL adds a new value to the cache with the given key and the provided ttl value.
@@ -157,5 +168,117 @@ func (l *tPool[K, V, C]) TTL(key K) (time.Duration, bool) {
 // Check [PoolTLRU.Upsert] on how Upsert works.
 func (l *tPool[K, V, C]) UpsertWithTTL(key K, value V, ttl time.Duration) (core.UpsertState, V) {
 	shard := l.mux(key)
-	return l.pool.shards[shard].UpsertWithTTL(key, value, ttl)
+	return l.shards[shard].UpsertWithTTL(key, value, ttl)
+}
+
+// Capacity returns the maximum allocated capacity of the LRU cache
+// across all sharded instances of [core.Shard].
+func (l *tPool[K, V, C]) Capacity() int {
+	return l.capacity
+}
+
+// Contains checks whether the key is present in the LRU cache.
+func (l *tPool[K, V, C]) Contains(key K) bool {
+	shard := l.mux(key)
+	return l.shards[shard].Contains(key)
+}
+
+// Delete removes the key from the cache and returns the evicted value.
+// It returns false if the key was not found in the cache.
+func (l *tPool[K, V, C]) Delete(key K) (V, bool) {
+	shard := l.mux(key)
+	value, ok := l.shards[shard].Delete(key)
+	if !ok {
+		return *new(V), false
+	}
+	return value, true
+}
+
+// Flush clears the LRU cache of all its keys and values across
+// all sharded instances.
+func (l *tPool[K, V, C]) Flush() {
+	for _, c := range l.shards {
+		c.Flush()
+	}
+}
+
+// Get retrieves the cache value using key.
+// It returns false if the key is not found.
+// It updates the key as 'recent' only in its respective shard.
+func (l *tPool[K, V, C]) Get(key K) (V, bool) {
+	shard := l.mux(key)
+	value, ok := l.shards[shard].Get(key)
+	if !ok {
+		return *new(V), false
+	}
+	return value, true
+}
+
+// Peek retrieves the cache value without updating it
+// to be the most recently used.
+// It returns false if the key is not found.
+func (l *tPool[K, V, C]) Peek(key K) (V, bool) {
+	shard := l.mux(key)
+	value, ok := l.shards[shard].Peek(key)
+	if !ok {
+		return *new(V), false
+	}
+	return value, true
+}
+
+// Put adds a new value to the cache with the given key.
+// It updates the key as 'recent' only in its respective shard.
+// It evicts the key only from the respective shard the key is linked to.
+func (l *tPool[K, V, C]) Put(key K, value V) {
+	shard := l.mux(key)
+	l.shards[shard].Put(key, value)
+}
+
+// ResetStats resets the stats of the sharded LRU cache.
+func (l *tPool[K, V, C]) ResetStats() {
+	for i := range l.shards {
+		l.shards[i].ResetStats()
+	}
+}
+
+// Shards returns the number of sharded instances in the LRU cache.
+func (l *tPool[K, V, C]) Shards() int {
+	return len(l.shards)
+}
+
+// Size returns the current size of the LRU cache
+// across all sharded instances.
+func (l *tPool[K, V, C]) Size() int {
+	size := 0
+	for i := range l.shards {
+		size += l.shards[i].Size()
+	}
+	return size
+}
+
+// Stats return the current stats of the sharded LRU cache.
+func (l *tPool[K, V, C]) Stats() core.Stats {
+	stats := core.Stats{}
+	for i := range l.shards {
+		st := l.shards[i].Stats()
+		stats.Hits += st.Hits
+		stats.Misses += st.Misses
+		stats.Evictions += st.Evictions
+		stats.Expirations += st.Expirations
+	}
+	return stats
+}
+
+// Upsert adds a new value to the cache with the given key.
+// It returns a value based on how the internal state of the cache changed.
+// It evicts or updates locally on the shard, instead of global cache.
+// Returns a value [core.UpsertState].
+//
+// It also returns a value based on [core.UpsertState]
+//   - [core.UpsertAddNoEviction] returns the zero value of V.
+//   - [core.UpsertAddWithEviction] returns the evicted value.
+//   - [core.UpsertReplace] returns the old value the key had.
+func (l *tPool[K, V, C]) Upsert(key K, value V) (core.UpsertState, V) {
+	shard := l.mux(key)
+	return l.shards[shard].Upsert(key, value)
 }